# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your API Documentation Title",
        "description": "This is the API documentation for the news aggregator service.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/preferences": {
            "get": {
                "tags": [
                    "Preferences"
                ],
                "summary": "Retrieve preferences for a specific user",
                "description": "Fetches the preferences for the specified user, including preferred news sources, categories, and authors.",
                "operationId": "86d82a29efa9769882266a99f89e8f6d",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "ID of the user to retrieve preferences for",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Preferences retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "preferences": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "user_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "preferred_source": {
                                                    "type": "string",
                                                    "example": "NewsAPI Org"
                                                },
                                                "preferred_category": {
                                                    "type": "string",
                                                    "example": ""
                                                },
                                                "preferred_author": {
                                                    "type": "string",
                                                    "example": "Nika Shakhnazarova"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-10-26T12:10:38.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "format": "date-time",
                                                    "example": "2024-10-26T12:22:59.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Preferences"
                ],
                "summary": "Set user preferences",
                "description": "Allows users to set their preferred news sources, categories, and authors.",
                "operationId": "a00b29a1fbbd8572ac8f659146ade2da",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "user_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "preferred_source": {
                                        "type": "string",
                                        "example": "NewsAPI Org"
                                    },
                                    "preferred_category": {
                                        "type": "string",
                                        "example": ""
                                    },
                                    "preferred_author": {
                                        "type": "string",
                                        "example": "Nika Shakhnazarova"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Preferences updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Preferences updated successfully"
                                        },
                                        "user_preferences": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        },
        "/api/personalized-feed": {
            "get": {
                "tags": [
                    "Preferences"
                ],
                "summary": "Fetch a personalized news feed",
                "description": "Retrieve a personalized news feed based on the user's preferences for source, category, and author.",
                "operationId": "69a4f37dd10424fbca9f28f187d71626",
                "parameters": [
                    {
                        "name": "user_id",
                        "in": "query",
                        "description": "ID of the user to retrieve the personalized feed for",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Personalized feed retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "‘Naked and Afraid’ star Sarah Danser dies in crash at age 34 - New York Post"
                                                    },
                                                    "content": {
                                                        "type": "string",
                                                        "example": "The reality TV star appeared on the survival series in 2017."
                                                    },
                                                    "author": {
                                                        "type": "string",
                                                        "example": "Nika Shakhnazarova"
                                                    },
                                                    "source": {
                                                        "type": "string",
                                                        "example": "NewsAPI Org"
                                                    },
                                                    "category": {
                                                        "type": "string",
                                                        "example": null,
                                                        "nullable": true
                                                    },
                                                    "published_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-10-25 08:07:00"
                                                    },
                                                    "image_url": {
                                                        "type": "string",
                                                        "example": "https://nypost.com/wp-content/uploads/sites/2/2024/10/92346424.jpg?quality=75&strip=all&w=1024"
                                                    },
                                                    "url": {
                                                        "type": "string",
                                                        "example": "https://nypost.com/2024/10/25/entertainment/naked-and-afraid-star-dies-in-crash-at-age-34/"
                                                    },
                                                    "created_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-10-26T09:21:35.000000Z"
                                                    },
                                                    "updated_at": {
                                                        "type": "string",
                                                        "format": "date-time",
                                                        "example": "2024-10-26T09:21:35.000000Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "first_page_url": {
                                            "type": "string",
                                            "example": "http://localhost:8000/api/personalized-feed?page=1"
                                        },
                                        "from": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "last_page_url": {
                                            "type": "string",
                                            "example": "http://localhost:8000/api/personalized-feed?page=1"
                                        },
                                        "links": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "url": {
                                                        "type": "string",
                                                        "example": null,
                                                        "nullable": true
                                                    },
                                                    "label": {
                                                        "type": "string",
                                                        "example": "&laquo; Previous"
                                                    },
                                                    "active": {
                                                        "type": "boolean",
                                                        "example": false
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "next_page_url": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "path": {
                                            "type": "string",
                                            "example": "http://localhost:8000/api/personalized-feed"
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "prev_page_url": {
                                            "type": "string",
                                            "example": null,
                                            "nullable": true
                                        },
                                        "to": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No preferences set for the user",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "No preferences set."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
...
}